chore: Update Angular project configuration and remove deprecated files
