+++ packages/doc-ai-angular/angular.json
@@ -43,7 +43,7 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "port": 8090
+            "port": 8099
           },
           "configurations": {
             "production": { "buildTarget": "doc-ai-angular:build:production" },
