+++ packages/doc-ai-angular/package.json
@@ -4,22 +4,20 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "concurrently -n ng,remoter \"ng serve\" \"vite --config vite.remoter.config.ts\"",
+    "dev": "concurrently -n ng,remoter \"ng serve\" \"pnpm -C remoter dev\"",
     "dev:ng": "ng serve",
-    "dev:remoter": "vite --config vite.remoter.config.ts",
+    "dev:remoter": "pnpm -C remoter dev",
     "build": "ng build",
-    "build:remoter": "vite build --config vite.remoter.config.ts",
+    "build:remoter": "pnpm -C remoter build",
     "preview": "ng serve --configuration production"
   },
   "dependencies": {
-    "@angular/common": "^19.0.0",
-    "zod": "catalog:",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/router": "^19.0.0",
-    "@opentiny/next-remoter": "workspace:*",
+    "@angular/common": "^21.2.0",
+    "@angular/compiler": "^21.2.0",
+    "@angular/core": "^21.2.0",
+    "@angular/forms": "^21.2.0",
+    "@angular/platform-browser": "^21.2.0",
+    "@angular/router": "^21.2.0",
     "@opentiny/next-sdk": "workspace:*",
     "rxjs": "^7.8.0",
     "tslib": "^2.6.0",
@@ -27,23 +25,14 @@
     "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.0",
-    "@angular/cli": "^19.0.0",
-    "@analogjs/vite-plugin-angular": "^2.3.1",
-    "@angular/build": "^19.2.22",
-    "@angular/compiler-cli": "^19.0.0",
-    "@opentiny/unplugin-tiny-vue": "catalog:",
-    "@opentiny/vue-vite-import": "catalog:",
-    "@vant/auto-import-resolver": "catalog:",
-    "@vitejs/plugin-vue": "catalog:",
+    "@angular-devkit/build-angular": "^21.2.0",
+    "@angular/cli": "^21.2.0",
+    "@angular/build": "^21.2.0",
+    "@angular/compiler-cli": "^21.2.0",
     "concurrently": "^9.1.0",
     "@types/node": "catalog:",
-    "unplugin-auto-import": "catalog:",
-    "unplugin-vue-components": "catalog:",
-    "vite-svg-loader": "catalog:",
     "dotenv": "catalog:",
     "sass": "^1.80.0",
-    "typescript": "~5.6.3",
-    "vite": "catalog:"
+    "typescript": "~5.9.0"
   }
 }
