+++ packages/doc-ai-angular/remoter/auto-imports.d.ts
