+++ packages/doc-ai-angular/remoter/index.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html lang="zh-CN">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" href="data:," />
+    <title>TinyRemoter - iframe host</title>
+    <style>
+      html, body { margin: 0; padding: 0; overflow: hidden; background: transparent; }
+    </style>
+  </head>
+  <body>
+    <div id="remoter-app"></div>
+    <script type="module" src="/src/main.ts"></script>
+  </body>
+</html>
