+++ packages/doc-ai-angular/remoter/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "doc-ai-angular-remoter",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@opentiny/next-remoter": "workspace:*",
+    "@opentiny/next-sdk": "workspace:*",
+    "vue": "catalog:"
+  },
+  "devDependencies": {
+    "@opentiny/unplugin-tiny-vue": "catalog:",
+    "@opentiny/vue-vite-import": "catalog:",
+    "@vant/auto-import-resolver": "catalog:",
+    "@vitejs/plugin-vue": "catalog:",
+    "unplugin-auto-import": "catalog:",
+    "unplugin-vue-components": "catalog:",
+    "vite": "catalog:",
+    "vite-svg-loader": "catalog:",
+    "typescript": "catalog:"
+  }
+}
