+++ packages/doc-ai-angular/remoter/src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
-import RemoterApp from './RemoterApp.vue'
+import App from './App.vue'
 
 // 创建并挂载 Vue 应用（仅包含 TinyRemoter），用于 iframe 内渲染
-const app = createApp(RemoterApp)
+const app = createApp(App)
 app.mount('#remoter-app')
