+++ packages/doc-ai-angular/remoter/src/skills/calculator/SKILL.md
