+++ packages/doc-ai-angular/remoter/src/skills/product-guide/SKILL.md
