+++ packages/doc-ai-angular/remoter/src/skills/product-guide/reference/product-shangjia.md
@@ -1,0 +1,60 @@
+---
+title: 商品上架
+impact: HIGH
+impactDescription: 商品上架是商品管理的重要环节，直接影响商品的销售与库存可见性，需规范流程以避免违规或信息缺失。
+type: operation-guide
+tags: [商品管理, 上架, 库存, 销售, 审核]
+---
+
+# 商品上架
+
+**Impact: HIGH** - 商品上架是商品从「待售/草稿」变为「在售、可被用户购买」的关键步骤。需确保信息完整、类目正确、库存与价格已配置，并符合平台规范后再上架。
+
+## 任务清单
+
+- [ ] 确认商品基础信息已填写完整（标题、主图、详情、类目）
+- [ ] 配置库存与价格（含促销价/会员价若需要）
+- [ ] 确认类目与属性符合平台规范，避免违规
+- [ ] 选择上架渠道/店铺（若有多渠道）
+- [ ] 执行上架操作并核对前台展示
+
+**何时需要「商品上架」说明：**
+
+- 用户询问如何把商品挂到店铺销售
+- 新商品创建完成后不知道下一步操作
+- 需要了解上架前必填项与审核要点
+- 区分「保存草稿」与「立即上架」的流程
+
+**与其它操作的边界：**
+
+- **商品创建**：填写信息、保存草稿，尚未进入销售
+- **商品上架**：将已就绪的商品发布为在售状态
+- **库存管理**：上架后的库存变更、预警、盘点
+
+## 基本流程
+
+1. **进入商品管理** → 找到待上架商品（草稿或待上架列表）。
+2. **补全必填项**：主图、标题、类目、价格、库存为常见必填；详情与属性按平台要求补全。
+3. **自检**：类目是否正确、是否有违禁词/违禁图、价格与库存是否合理。
+4. **提交上架**：点击「上架」或「提交审核」（视平台是否需审核而定）。
+5. **验证**：在前台或店铺端确认商品已展示、可加购。
+
+## 常见场景
+
+### 单个商品上架
+
+适用于已编辑好的单个商品，检查无误后直接上架。注意上架后若修改关键信息（如类目、价格），可能触发重新审核或限流。
+
+### 批量上架
+
+适用于活动前或导入后的批量发布。建议先小批量试跑，确认类目与模板无误后再全量上架，避免批量违规。
+
+### 定时上架
+
+若平台支持「定时上架」，可提前配置好商品与上架时间，适合活动准时开卖或分批上架策略。
+
+## 注意事项
+
+- **审核型平台**：上架后可能为「待审核」状态，通过后才在前台展示；需预留审核时间。
+- **库存为 0**：部分平台不允许库存为 0 时上架，需先设库存再上架。
+- **类目与属性**：类目选错可能导致下架或处罚，上架前务必核对。
