+++ packages/doc-ai-angular/src/app/app.component.scss
@@ -1,3 +1,4 @@
+// 根布局：左侧主内容 70%，右侧 AI 助手固定 30%
 .app-container {
   position: relative;
   width: 100%;
@@ -5,26 +6,26 @@
 }
 
 .main-content {
-  width: 100%;
+  width: 70%;
   min-height: 100vh;
+  overflow: auto;
 }
 
-/**
- * TinyRemoter iframe 定位策略：
- *
- * 与 Intercom/Crisp 等真实聊天组件相同的做法——将 iframe 固定在右下角。
- * 优点：主内容区域完全可交互，不需要 pointer-events 穿透等复杂机制。
- * 约定：TinyRemoter 浮动按钮位于右下角，聊天面板向上/左展开，
- *       均在 iframe 区域（420x700）内渲染。
- */
-.remoter-frame {
+.remoter-sidebar {
   position: fixed;
-  bottom: 0;
+  top: 0;
   right: 0;
-  width: 420px;
-  height: 700px;
+  width: 30%;
+  height: 100vh;
+  border-left: 1px solid #eee;
+  background: #fff;
+  z-index: 10;
+}
+
+.remoter-frame {
+  width: 100%;
+  height: 100%;
+  min-height: 100vh;
   border: none;
-  background: transparent;
-  z-index: 9999;
-  pointer-events: auto;
+  display: block;
 }
